+++ test/time_ago_provider_test.dart
@@ -1,14 +1,25 @@
 import 'package:time_ago_provider/time_ago_provider.dart';
 
-void main()
-{
+void main() {
   //Normal Usage
-  int myTimeStamp = DateTime.now().millisecondsSinceEpoch; // Get the current time stamp
-  String timeAgo = TimeAgo.getTimeAgo(timeStamp: myTimeStamp); //getTimeAgo(int timeStamp) returns a String " X minutes/hours/days/months/years ago ..."
+  int myTimeStamp =
+      DateTime.now().millisecondsSinceEpoch; // Get the current time stamp
+  String timeAgo = TimeAgo.getTimeAgo(
+      timeStamp:
+          myTimeStamp); //getTimeAgo(int timeStamp) returns a String " X minutes/hours/days/months/years ago ..."
   print("Posted $timeAgo");
 
   //With Language
   int mySecondTimeStamp = DateTime.now().millisecondsSinceEpoch;
-  String germanTimeAgo = TimeAgo.getTimeAgo(timeStamp: mySecondTimeStamp, language: Language.GERMAN); //Default Language ENGLISH
+  String germanTimeAgo = TimeAgo.getTimeAgo(
+      timeStamp: mySecondTimeStamp,
+      language: Language.GERMAN); //Default Language ENGLISH
   print("Gepostet $germanTimeAgo");
+
+  //In Nepali
+  int myThirdTimeStamp = DateTime.now().millisecondsSinceEpoch;
+  String nepaliTimeAgo = TimeAgo.getTimeAgo(
+      timeStamp: mySecondTimeStamp,
+      language: Language.NEPALI); //Default Language ENGLISH
+  print("$nepaliTimeAgo प्रकाशित");
 }