+++ CHANGELOG.md
@@ -1,7 +1,14 @@
-## 0.0.1
+## 0.1.0+1
 
-- Initial version.
+- Support for Nepali Language added.
+- Fixed some mistranslation for Hindi Language.
 
 ## 0.1.0
 
 - Stable version.
+
+## 0.0.1
+
+- Initial version.
+
+